+++ CHANGELOG.md
@@ -35,6 +35,10 @@ These were there since 0.1.0 but I forgot to export the `path` module so I'm now
 
 - added `getFirstAndLast`
 
+### Array -> Number
+
+- added `arraySum`
+
 ### Array -> (Number -> Number)
 
 - added `makePolynomial`